Final Project task complete
CMS using laravel

+++ cms-laravel-app/app/Http/Controllers/CategoriesController.php
@@ -2,10 +2,10 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
 use App\Category;
+use Illuminate\Http\Request;
+use App\Http\Requests\Categories\CreateCategoryRequest;
 use App\Http\Requests\Categories\UpdateCategoriesRequest;
-use App\Http\Requests\Categories\CreateCategoriesRequest;
 
 class CategoriesController extends Controller
 {
@@ -16,7 +16,6 @@ class CategoriesController extends Controller
      */
     public function index()
     {
-
         return view('categories.index')->with('categories', Category::all());
     }
 
@@ -36,13 +35,14 @@ class CategoriesController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(CreateCategoriesRequest $request)
+    public function store(CreateCategoryRequest $request)
     {
         Category::create([
-            'name' => $request->name
+          'name' => $request->name
         ]);
 
         session()->flash('success', 'Category created successfully.');
+
         return redirect(route('categories.index'));
     }
 
@@ -78,9 +78,11 @@ class CategoriesController extends Controller
     public function update(UpdateCategoriesRequest $request, Category $category)
     {
         $category->update([
-            'name' => $request->name
+          'name' => $request->name
         ]);
+
         session()->flash('success', 'Category updated successfully.');
+
         return redirect(route('categories.index'));
     }
 
@@ -92,8 +94,16 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+        if ($category->posts->count() > 0) {
+          session()->flash('error', 'Category cannot be deleted because it has some posts.');
+
+          return redirect()->back();
+        }
+        
         $category->delete();
-        session()->flash('success', 'Category deleted successfully');
+
+        session()->flash('success', 'Category deleted successfully.');
+
         return redirect(route('categories.index'));
     }
 }
